Group DOM functions and add dummy data
